+++ hlslang/MachineIndependent/preprocessor/mojoshader_lexer.cpp
@@ -1196,6 +1196,14 @@ yy192:
 	{
                         if (s->report_whitespace)
                             RET(' ');
+						// Microsoft's preprocessor allows multiline comments
+						//  before a preprocessor directive, even though C/C++
+						//  doesn't. See if we've hit this case.
+						if (s->tokenval == ((Token) '\n'))  // was start of line?
+						{
+							update_state(s, eoi, cursor, token, (Token) '\n');
+							goto ppdirective;  // may jump back to scanner_loop.
+						}
                         goto scanner_loop;
                     }
 }